Minor trolling

+++ app/src/main/java/app/model/credentials_info/Manager.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class Manager extends Staff{
     public Manager(String email, String password,
                    String name, Date startingDate, String position, float salary,
-                   long purchasesRegistered, int sessionsCompleted, long[] sessionsDone) {
+                   int purchasesRegistered, int sessionsCompleted, int[] sessionsDone) {
         super(email, password, 
               name, startingDate, position, salary, 
               purchasesRegistered, sessionsCompleted, sessionsDone);
